remove window main

+++ src/SystemIcon.cpp
@@ -73,7 +73,7 @@ sk_sp<SkImage> SystemIcon::iconToImg(HICON icon, const int& size)
 	DrawIconEx(hdcMemory, 0, 0, icon, size, size, 0, NULL, DI_NORMAL);
 	SkImageInfo imgInfo = SkImageInfo::MakeN32Premul(size, size);
 	std::vector<int32_t> pixSrcData;
-	double dataSize = size * 4 * size;
+	unsigned long dataSize = size * 4 * size;
 	pixSrcData.resize(dataSize);
 	SkPixmap pixmap(imgInfo, &pixSrcData.front(), size * 4);
 	BITMAPINFO info = { sizeof(BITMAPINFOHEADER), size, 0 - size, 1, 32, BI_RGB, dataSize, 0, 0, 0, 0 };
